+++ src/lib/components/custom/sidebar/index.ts
@@ -3,3 +3,5 @@ export { default as SidebarHeader } from './sidebar-header.svelte';
 export { default as SidebarFooter } from './sidebar-footer.svelte';
 export { default as SidebarTree } from './sidebar-tree.svelte';
 export { default as AddPanelAndViewTrigger } from './add-panel-and-view-trigger.svelte';
+export { default as UpdatePanelDialog } from './update-panel-dialog.svelte';
+export { default as UpdateLoadDialog } from './update-load-dialog.svelte';