+++ src/lib/components/custom/table/(components)/data-table-add-load.svelte
@@ -4,12 +4,12 @@
 	import { GenericPhaseMainLoadForm } from '@/components/custom/load';
 	import { CirclePlus } from '@/assets/icons';
 	import type { SuperValidated } from 'sveltekit-superforms';
-	import { type PhaseMainLoadSchema } from '@/schema/load';
+	import { type GenericPhaseMainLoadSchema } from '@/schema/load';
 	import type { HighestUnitSchema } from '@/schema';
 	import { cn } from '@/utils';
 
 	interface Props {
-		phase_main_load_form: SuperValidated<PhaseMainLoadSchema>;
+		phase_main_load_form: SuperValidated<GenericPhaseMainLoadSchema>;
 		highest_unit: HighestUnitSchema;
 	}